adminpenkit/modules/advanced_network.py
import nmap
import socket
import requests
import threading
from modules.base_module import BaseModule

class AdvancedNetworkTools(BaseModule):
    def __init__(self):
        super().__init__()
        self.name = "Advanced Network Tools"
        
    def initialize(self):
        return True
        
    def execute(self):
        return self.get_network_status()
        
    def cleanup(self):
        return True
        
    def get_network_status(self):
        return {
            "topology": self.map_network_topology(),
            "traffic": self.analyze_traffic(),
            "protocols": self.detect_protocols()
        }
        
    def map_network_topology(self):
        # Network mapping implementation
        pass
        
    def analyze_traffic(self):
        # Traffic analysis implementation
        pass
        
    def detect_protocols(self):
        # Protocol detection implementation
        pass
