requirements/requirements_base.txt
psutil>=5.9.0
python-nmap>=0.7.1
paramiko>=2.11.0
requests>=2.28.0
matplotlib>=3.8.0
pytest>=7.3.0